+++ tests/integration/cases/call_per_element_false/ObjectiveC_call.m
@@ -1,0 +1,6 @@
+#import <Foundation/Foundation.h>
+#pragma clang diagnostic ignored "-Wdeprecated-non-prototype"
+void process();
+void check_(void) {
+process(1);
+}