+++ tests/integration/cases/call_transform_no_wrapper/ObjectiveC_call.m
@@ -1,0 +1,8 @@
+#import <Foundation/Foundation.h>
+#pragma clang diagnostic ignored "-Wdeprecated-non-prototype"
+id process();
+void check_(void) {
+process(@"hello");
+process(42);
+process(@YES);
+}